Move CalibrationEntry and CalibrationPublisher to dedicated file for future deprecation of instruction schedule map.
